restructuring files into new model package
