+++ src/main/java/com/example/courseapi/controller/CourseController.java
@@ -1,5 +1,7 @@
 package com.example.courseapi.controller;
 
+import com.example.courseapi.model.Course;
+import com.example.courseapi.model.Topic;
 import com.example.courseapi.services.CourseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
