+++ src/main/java/com/example/courseapi/model/Course.java
@@ -1,27 +1,36 @@
-package com.example.courseapi.controller;
+package com.example.courseapi.model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 
 @Entity
-public class Topic {
+public class Course {
     @Getter
     @Id
     private String id;
     private String  name;
     private String description;
 
+
+
+    @ManyToOne
+    @JoinColumn(name = "topic_id")
+    private Topic topic;
+
     //This is a no arg constructor
-    public Topic() {
+    public Course() {
 
     }
 
     //This is a constructor that initializes the objects
-    public Topic(String id, String name, String description) {
+    public Course(String id, String name, String description, String topicId) {
         this.id = id;
         this.name = name;
         this.description = description;
+        this.topic = new Topic(topicId, "","");
     }
 
     public void setId(String id) {
@@ -42,6 +51,14 @@ public class Topic {
 
     public void setDescription(String description) {
         this.description = description;
+    }
+
+    public Topic getTopic() {
+        return topic;
+    }
+
+    public void setTopic(Topic topic) {
+        this.topic = topic;
     }
