+++ src/main/java/com/example/courseapi/model/Topic.java
@@ -1,36 +1,27 @@
-package com.example.courseapi.controller;
+package com.example.courseapi.model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 
 @Entity
-public class Course {
+public class Topic {
     @Getter
     @Id
     private String id;
     private String  name;
     private String description;
 
-
-
-    @ManyToOne
-    @JoinColumn(name = "topic_id")
-    private Topic topic;
-
     //This is a no arg constructor
-    public Course() {
+    public Topic() {
 
     }
 
     //This is a constructor that initializes the objects
-    public Course(String id, String name, String description, String topicId) {
+    public Topic(String id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
-        this.topic = new Topic(topicId, "","");
     }
 
     public void setId(String id) {
@@ -51,14 +42,6 @@ public class Course {
 
     public void setDescription(String description) {
         this.description = description;
-    }
-
-    public Topic getTopic() {
-        return topic;
-    }
-
-    public void setTopic(Topic topic) {
-        this.topic = topic;
     }
