+++ src/main/java/com/example/courseapi/repository/TopicRepository.java
@@ -1,7 +1,6 @@
 package com.example.courseapi.repository;
 
-import com.example.courseapi.controller.Course;
-import com.example.courseapi.controller.Topic;
+import com.example.courseapi.model.Topic;
 import org.springframework.data.repository.CrudRepository;
 
 public interface TopicRepository extends CrudRepository<Topic, String> {
